refactor: separate responsibility
