+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/school/SchoolReaderImpl.kt
@@ -2,27 +2,18 @@ package com.sexysisters.tojserverv2.infrastructure.school
 
 import com.sexysisters.tojserverv2.domain.school.School
 import com.sexysisters.tojserverv2.domain.school.design.SchoolReader
-import com.sexysisters.tojserverv2.domain.school.type.getDivision
-import com.sexysisters.tojserverv2.domain.school.type.getKind
-import com.sexysisters.tojserverv2.infrastructure.neis.SchoolInfoReader
+import com.sexysisters.tojserverv2.domain.school.exception.SchoolExcpetion
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 
 @Component
 class SchoolReaderImpl(
-    private val schoolInfoReader: SchoolInfoReader,
+    private val schoolRepository: SchoolRepository,
 ) : SchoolReader {
 
     @Transactional
     override fun findSchoolByCode(code: String): School {
-        val schoolInfoResponse = schoolInfoReader.findSchoolByCode(code)
-        val school = schoolInfoResponse.toEntity()
-
-        val kind = schoolInfoResponse.kind
-        school.kind = getKind(kind)
-        val division = schoolInfoResponse.division
-        school.division = getDivision(division)
-
-        return school
+        return schoolRepository.findByCode(code)
+            ?: throw SchoolExcpetion.SchoolNotFound()
     }
 }
