+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/school/SchoolRepository.kt
@@ -5,4 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 
 interface SchoolRepository : JpaRepository<School, Long> {
     fun existsByCode(code: String): Boolean
+    fun findByCode(code: String): School?
 }
